refactor(SDK-2575): update package and class names as per the naming conventions
