+++ android/src/main/java/com/clevertap/rnsignedcallandroid/ClevertapSignedCallPackage.kt
@@ -1,14 +1,13 @@
-package com.clevertapsignedcallreactnative
+package com.clevertap.rnsignedcallandroid
 
 import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
 
-
-class ClevertapSignedCallReactNativePackage : ReactPackage {
+class ClevertapSignedCallPackage : ReactPackage {
   override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
-    return listOf(ClevertapSignedCallReactNativeModule(reactContext))
+    return listOf(ClevertapSignedCallModule(reactContext))
   }
 
   override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {
