+++ example/android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.clevertapsignedcallreactnativeexample",
+    package = "com.example.rnsignedcall",
 )
 
 android_resource(
     name = "res",
-    package = "com.clevertapsignedcallreactnativeexample",
+    package = "com.example.rnsignedcall",
     res = "src/main/res",
 )
