+++ example/android/app/src/main/java/com/example/rnsignedcall/MainApplication.java
@@ -1,4 +1,4 @@
-package com.clevertapsignedcallreactnativeexample;
+package com.example.rnsignedcall;
 
 import android.app.Application;
 import android.content.Context;
@@ -9,7 +9,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.config.ReactFeatureFlags;
 import com.facebook.soloader.SoLoader;
-import com.clevertapsignedcallreactnativeexample.newarchitecture.MainApplicationReactNativeHost;
+import com.example.rnsignedcall.newarchitecture.MainApplicationReactNativeHost;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
